+++ Godeps/_workspace/src/github.com/restic/chunker/chunker.go
