+++ Godeps/_workspace/src/github.com/restic/chunker/doc.go
