+++ Godeps/_workspace/src/github.com/restic/chunker/polynomials.go
