+++ Godeps/_workspace/src/github.com/restic/chunker/polynomials_test.go
@@ -4,7 +4,7 @@ import (
 	"strconv"
 	"testing"
 
-	"github.com/restic/restic/chunker"
+	"github.com/restic/chunker"
 	. "github.com/restic/restic/test"
 )
