+++ archiver.go
@@ -10,8 +10,8 @@ import (
 	"sort"
 	"sync"
 
+	"github.com/restic/chunker"
 	"github.com/restic/restic/backend"
-	"github.com/restic/restic/chunker"
 	"github.com/restic/restic/debug"
 	"github.com/restic/restic/pack"
 	"github.com/restic/restic/pipe"
