+++ crypto/crypto_test.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/restic/restic/chunker"
+	"github.com/restic/chunker"
 	"github.com/restic/restic/crypto"
 	. "github.com/restic/restic/test"
 )
